// types/datePeriod.ts
export enum DatePeriod {
  LAST_SEVEN_DAYS = "LAST_SEVEN_DAYS", // eslint-disable-line no-unused-vars
  LAST_FOURTEEN_DAYS = "LAST_FOURTEEN_DAYS", // eslint-disable-line no-unused-vars
  LAST_THIRTY_DAYS = "LAST_THIRTY_DAYS", // eslint-disable-line no-unused-vars
  LAST_NINETY_DAYS = "LAST_NINETY_DAYS", // eslint-disable-line no-unused-vars
}